Update database.pp
First of all - great module, very useful!

We are using puppet enterprise in an environment where passwords are not generally known by all team members. When running the puppet agent to install oracle, the database.pp causes the system passwords to be output in plain-text to the puppet console. In our environment this discloses information to people who do not need to know the passwords but do need to view server status and information on the puppet console.
The most important part of this is the "show_diff" setting which prevents the console output.
For 'replace=>false': We have later on added code to copy /dev/null over the response file so that it is not left on the server holding the plain-text passwords. I'm not sure if this was the best fix but we needed something that didn't cause the file to be recreated on the next puppet run.